+++ src/main/java/com/almostreliable/energymeter/block/entity/MeterBlockEntity.java
@@ -124,15 +124,19 @@ public class MeterBlockEntity extends BlockEntity implements TickableMenuBlockEn
     @Override
     public void setChanged() {
         super.setChanged();
-        for (var observer : observers) {
-            observer.onChange(this);
+        synchronized (observers) {
+            for (var observer : observers) {
+                observer.onChange(this);
+            }
         }
     }
 
     @Override
     public void setRemoved() {
-        for (var observer : observers) {
-            observer.onRemove(this);
+        synchronized (observers) {
+            for (var observer : observers) {
+                observer.onRemove(this);
+            }
         }
         super.setRemoved();
     }
@@ -140,7 +144,7 @@ public class MeterBlockEntity extends BlockEntity implements TickableMenuBlockEn
     @Override
     public void tick(ServerLevel level) {
         if (!transferMode.isCorrectlyConfigured(ioConfig::hasInput, ioConfig::hasOutput)) {
-            connectionStatus = ConnectionStatus.DISCONNECTED;
+            updateConnectionStatus();
             return;
         }
 
@@ -151,7 +155,7 @@ public class MeterBlockEntity extends BlockEntity implements TickableMenuBlockEn
             onIntervalReached(level);
         }
 
-        connectionStatus = energyRate > 0 ? transferMode.activeStatus : ConnectionStatus.IDLE;
+        updateConnectionStatus();
     }
 
     private void onIntervalReached(ServerLevel level) {
@@ -175,7 +179,9 @@ public class MeterBlockEntity extends BlockEntity implements TickableMenuBlockEn
         totalEnergy += measuredEnergy.total();
 
         var energyRateChanged = lastEnergyRate != energyRate;
-        if (energyRateChanged || lastTotalEnergy != totalEnergy) {
+        var statusChanged = updateConnectionStatus();
+
+        if (!statusChanged && (energyRateChanged || lastTotalEnergy != totalEnergy)) {
             setChanged();
         }
 
@@ -193,12 +199,31 @@ public class MeterBlockEntity extends BlockEntity implements TickableMenuBlockEn
 
     private void onConnectionRelevantSettingChanged() {
         if (!(level instanceof ServerLevel serverLevel)) return;
+
         serverLevel.invalidateCapabilities(worldPosition);
         serverLevel.updateNeighborsAt(worldPosition, getBlockState().getBlock());
         energyHandler.clearOutputCacheAndReset();
         energyRate = 0;
         syncEnergyRate(serverLevel);
+
+        if (!updateConnectionStatus()) setChanged();
+    }
+
+    private boolean updateConnectionStatus() {
+        var newStatus = ConnectionStatus.IDLE;
+
+        if (transferMode.isCorrectlyConfigured(ioConfig::hasInput, ioConfig::hasOutput)) {
+            newStatus = energyRate > 0 ? transferMode.activeStatus : ConnectionStatus.IDLE;
+        } else {
+            newStatus = ConnectionStatus.DISCONNECTED;
+        }
+
+        if (connectionStatus == newStatus) return false;
+
+        connectionStatus = newStatus;
         setChanged();
+
+        return true;
     }
 
     // fall-back to clear output capability cache in case a block doesn't invalidate the capability
@@ -252,7 +277,6 @@ public class MeterBlockEntity extends BlockEntity implements TickableMenuBlockEn
     public void setTransferMode(TransferMode transferMode) {
         this.transferMode = transferMode;
         onConnectionRelevantSettingChanged();
-        setChanged();
     }
 
     public MeasureMode getMeasureMode() {