+++ src/main/java/com/almostreliable/energymeter/compat/cct/PeripheralCapabilityAdapter.java
@@ -14,16 +14,6 @@ import org.jspecify.annotations.Nullable;
 
 class PeripheralCapabilityAdapter implements ICapabilityAdapter<MeterPeripheral> {
 
-    private static final Identifier PERIPHERAL_ID = Identifier.fromNamespaceAndPath(
-        ComputerCraftAPI.MOD_ID,
-        "peripheral"
-    );
-    static final BlockCapability<IPeripheral, Direction> PERIPHERAL_CAPABILITY = BlockCapability.create(
-        PERIPHERAL_ID,
-        IPeripheral.class,
-        Direction.class
-    );
-
     private final MeterPeripheral peripheral;
 
     PeripheralCapabilityAdapter(MeterBlockEntity entity) {